removed separate file for GeoExtent
