+++ city_metrix/layers/albedo.py
@@ -2,9 +2,8 @@ import ee
 import xarray
 from dask.diagnostics import ProgressBar
 
-from .layer import (Layer, get_image_collection, set_resampling_for_continuous_raster,
+from .layer import (GeoExtent, Layer, get_image_collection, set_resampling_for_continuous_raster,
                     validate_raster_resampling_method)
-from .layer_geometry import GeoExtent
 
 DEFAULT_SPATIAL_RESOLUTION = 10
 DEFAULT_RESAMPLING_METHOD = 'bilinear'
