+++ city_metrix/layers/alos_dsm.py
@@ -2,8 +2,7 @@ import ee
 import xee
 import xarray as xr
 
-from .layer import Layer, get_image_collection, set_resampling_for_continuous_raster, validate_raster_resampling_method
-from .layer_geometry import GeoExtent
+from .layer import GeoExtent, Layer, get_image_collection, set_resampling_for_continuous_raster, validate_raster_resampling_method
 
 DEFAULT_SPATIAL_RESOLUTION = 30
 DEFAULT_RESAMPLING_METHOD = 'bilinear'
