+++ city_metrix/layers/average_net_building_height.py
@@ -3,8 +3,7 @@ import xarray as xr
 import xee
 import ee
 
-from .layer import Layer, get_image_collection
-from .layer_geometry import GeoExtent
+from .layer import GeoExtent, Layer, get_image_collection
 
 DEFAULT_SPATIAL_RESOLUTION = 100
