+++ city_metrix/layers/era_5_hottest_day.py
@@ -9,8 +9,7 @@ import xarray as xr
 import glob
 
 from tests.resources.layer_dumps.conftest import sample_aoi
-from .layer import Layer
-from .layer_geometry import GeoExtent
+from .layer import GeoExtent, Layer
 
 
 class Era5HottestDay(Layer):
