+++ city_metrix/layers/height_above_nearest_drainage.py
@@ -2,8 +2,7 @@ from dask.diagnostics import ProgressBar
 import xarray as xr
 import ee
 
-from .layer import Layer, get_image_collection
-from .layer_geometry import GeoExtent
+from .layer import GeoExtent, Layer, get_image_collection
 
 DEFAULT_SPATIAL_RESOLUTION = 30
