+++ city_metrix/layers/ndvi_sentinel2_gee.py
@@ -1,7 +1,6 @@
 import ee
 
-from .layer import Layer, get_image_collection
-from .layer_geometry import GeoExtent
+from .layer import GeoExtent, Layer, get_image_collection
 
 DEFAULT_SPATIAL_RESOLUTION = 10
