+++ city_metrix/layers/tree_cover.py
@@ -1,11 +1,10 @@
-from .layer import Layer, get_image_collection
+from .layer import GeoExtent, Layer, get_image_collection
 
 from dask.diagnostics import ProgressBar
 import xarray as xr
 import xee
 import ee
 
-from .layer_geometry import GeoExtent
 
 DEFAULT_SPATIAL_RESOLUTION = 10
